Implemented code review feedback
