+++ XamarinCommunityToolkit/Views/AvatarView/AvatarView.shared.cs
@@ -27,7 +27,7 @@ namespace Xamarin.CommunityToolkit.UI.Views
 
 		public static readonly BindableProperty FontAttributesProperty = BindableProperty.Create(nameof(FontAttributes), typeof(FontAttributes), typeof(AvatarView), FontAttributes.None, propertyChanged: OnValuePropertyChanged);
 
-		public static readonly BindableProperty ColorThemeProperty = BindableProperty.Create(nameof(ColorTheme), typeof(IColorTheme), typeof(AvatarView), ColorThemes.Default, propertyChanged: OnValuePropertyChanged);
+		public static readonly BindableProperty ColorThemeProperty = BindableProperty.Create(nameof(ColorTheme), typeof(IColorTheme), typeof(AvatarView), Views.ColorTheme.Default, propertyChanged: OnValuePropertyChanged);
 
 		public double Size
 		{
