+++ XamarinCommunityToolkit/Views/AvatarView/IColorTheme.shared.cs
@@ -5,10 +5,6 @@ namespace Xamarin.CommunityToolkit.UI.Views
 {
 	public interface IColorTheme
 	{
-		Color[] BackgroundColors { get; set; }
-
-		Color[] ForegroundColors { get; set; }
-
 		Color GetForegroundColor(string text);
 
 		Color GetBackgroundColor(string text);
