+++ XamarinCommunityToolkitSample/Pages/Views/AvatarViewPage.xaml
@@ -22,7 +22,7 @@
 
                             <views:AvatarView Size="{Binding Value, Source={x:Reference Slider}}"
                                               Text="{Binding Initials}" Source="{Binding Source}"
-                                              ColorTheme="{x:Static views:ColorThemes.Default}"/>
+                                              ColorTheme="{x:Static views:ColorTheme.Jungle}"/>
 
                             <Label Text="{Binding Name}" FontSize="20" VerticalOptions="CenterAndExpand" />
                         </StackLayout>
